glue

+++ run_classifier.py
@@ -740,7 +740,7 @@ def file_based_input_fn_builder(input_file, seq_length, is_training,
       "input_ids": tf.io.FixedLenFeature([seq_length], tf.int64),
       "input_mask": tf.io.FixedLenFeature([seq_length], tf.int64),
       "segment_ids": tf.io.FixedLenFeature([seq_length], tf.int64),
-      "label_ids": tf.io.FixedLenFeature([], tf.int64),
+      "label_ids": tf.io.FixedLenFeature([], tf.float32),
       "is_real_example": tf.io.FixedLenFeature([], tf.int64),
   }
